+++ tests/index_generator/test_np_index_generator.py
@@ -13,13 +13,13 @@ from connectlib.index_generator.np_index_generator import NpIndexGenerator
 def test_np_index_generator_drop_last(batch_size):
     # Check that if the last batch is dropped, all batches always have the same size, i.e. the batch size.
     nig = NpIndexGenerator(
-        n_samples=10,
         batch_size=batch_size,
         shuffle=True,
         drop_last=True,
         seed=42,
         num_updates=20,
     )
+    nig.n_samples = 10
     for _ in range(9):
         assert len(nig.__next__()) == batch_size, "All batches do not have the same size when drop_last is set to True."
 
@@ -30,14 +30,13 @@ def test_np_index_generator_keep_last(n_samples, batch_size):
     # check that the first n-1's are of size : batch size and that the nth one is of size n_samples%batch_size.
 
     nig = NpIndexGenerator(
-        n_samples=n_samples,
         batch_size=batch_size,
         shuffle=True,
         drop_last=False,
         seed=42,
         num_updates=2000,
     )
-
+    nig.n_samples = n_samples
     n_batch_per_epoch = nig._n_batch_per_epoch
     last_batch_size = n_samples % batch_size
 
@@ -65,29 +64,28 @@ def test_np_index_generator_100_epoch(n_samples, batch_size, drop_last):
     # The generator has also been tested locally to generates 10 000 epochs.
 
     nig = NpIndexGenerator(
-        n_samples=n_samples,
         batch_size=batch_size,
         shuffle=True,
         drop_last=drop_last,
         seed=42,
         num_updates=2000,
     )
+    nig.n_samples = n_samples
     for _ in range(100 * nig._n_batch_per_epoch):
         assert len(nig.__next__()) > 0
 
 
 def test_np_index_generator_n_samples_negative():
     # n_samples can't be negative
-
+    nig = NpIndexGenerator(
+        batch_size=12,
+        shuffle=True,
+        drop_last=True,
+        seed=42,
+        num_updates=40,
+    )
     with pytest.raises(ValueError):
-        _ = NpIndexGenerator(
-            n_samples=-4,
-            batch_size=12,
-            shuffle=True,
-            drop_last=True,
-            seed=42,
-            num_updates=40,
-        )
+        nig.n_samples = -4
 
 
 def test_np_index_generator_batch_size_negative():
@@ -95,7 +93,6 @@ def test_np_index_generator_batch_size_negative():
 
     with pytest.raises(ValueError):
         _ = NpIndexGenerator(
-            n_samples=4,
             batch_size=-12,
             shuffle=True,
             drop_last=True,
@@ -113,14 +110,16 @@ def test_np_index_generator_batch_size_negative():
 )
 def test_np_index_generator_batch_shuffle(seed, results):
     # Check that the shuffling is properly seeded.
-    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=seed, num_updates=40)
+    nig = NpIndexGenerator(batch_size=3, shuffle=True, drop_last=True, seed=seed, num_updates=40)
+    nig.n_samples = 3
     for result in results:
         assert np.array_equal(result, nig.__next__())
 
 
 def test_np_index_generator_not_batch_shuffle():
     # Check that we can have not shuffled batches.
-    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=False, drop_last=True, seed=42, num_updates=40)
+    nig = NpIndexGenerator(batch_size=3, shuffle=False, drop_last=True, seed=42, num_updates=40)
+    nig.n_samples = 3
     res = np.arange(3)
 
     for _ in range(10):
@@ -134,13 +133,13 @@ def test_np_index_generator_not_batch_shuffle():
 def test_np_index_generator_check_epoch_consitency(n_samples, batch_size, drop_last):
     # Check that for each epoch, all indexes are not duplicated
     nig = NpIndexGenerator(
-        n_samples=n_samples,
         batch_size=batch_size,
         shuffle=True,
         drop_last=drop_last,
         seed=42,
         num_updates=40,
     )
+    nig.n_samples = n_samples
 
     n_epoch = nig.n_epoch_generated
     generated_epoch_indexes = np.array([])
@@ -163,13 +162,13 @@ def test_np_index_generator_check_epoch_consitency(n_samples, batch_size, drop_l
 def test_np_index_generator_count_batch_per_epoch(n_samples, batch_size, drop_last, expected):
     # Check that for each epoch, the right number of batch is generated
     nig = NpIndexGenerator(
-        n_samples=n_samples,
         batch_size=batch_size,
         shuffle=True,
         drop_last=drop_last,
         seed=42,
         num_updates=50,
     )
+    nig.n_samples = n_samples
     n_epoch = 0
 
     for i in range(1, 30):
@@ -184,7 +183,8 @@ def test_np_index_generator_statefulness(session_dir):
     # Check that after saving and loading, the batch generator still has the same outputs.
     indexer_path = Path(tempfile.mkdtemp(dir=session_dir)) / "my_indexer"
 
-    nig = NpIndexGenerator(n_samples=5, batch_size=3, shuffle=True, drop_last=True, seed=42, num_updates=50)
+    nig = NpIndexGenerator(batch_size=3, shuffle=True, drop_last=True, seed=42, num_updates=50)
+    nig.n_samples = 5
 
     # Artificial 10 first batch
     for _ in range(10):
@@ -207,7 +207,8 @@ def test_np_index_generator_statefulness(session_dir):
 
 def test_np_index_generator_num_updates():
     """Check that the iterator stops after num_updates iterations"""
-    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=10)
+    nig = NpIndexGenerator(batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=10)
+    nig.n_samples = 3
     for _ in range(10):
         next(nig)
     with pytest.raises(StopIteration):
@@ -216,7 +217,8 @@ def test_np_index_generator_num_updates():
 
 def test_np_index_generator_counter():
     """Check that the counter is incremented at each call"""
-    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=10)
+    nig = NpIndexGenerator(batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=10)
+    nig.n_samples = 3
     assert nig.counter == 0
     next(nig)
     assert nig.counter == 1
@@ -225,7 +227,8 @@ def test_np_index_generator_counter():
 def test_np_index_generator_reset():
     """Test the reset function"""
     num_updates = 10
-    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates)
+    nig = NpIndexGenerator(batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates)
+    nig.n_samples = 3
     for _ in range(num_updates):
         next(nig)
 
@@ -242,12 +245,10 @@ def test_np_index_generator_torch_dataloader(num_workers):
     """Test that the index generator works with a torch dataloader"""
     n_samples = 100
     num_updates = 10
-    nig = NpIndexGenerator(
-        n_samples=n_samples, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates
-    )
-    nig_copy = NpIndexGenerator(
-        n_samples=n_samples, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates
-    )
+    nig = NpIndexGenerator(batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates)
+    nig.n_samples = n_samples
+    nig_copy = NpIndexGenerator(batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates)
+    nig_copy.n_samples = n_samples
 
     class DummyDataset(torch.utils.data.Dataset):
         def __init__(self):